Moved Data Classes to Model package
