+++ composeApp/src/commonMain/kotlin/database/model/Event.kt
@@ -1,0 +1,14 @@
+package database.model
+
+import androidx.compose.runtime.LaunchedEffect
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity
+data class Event (
+    @PrimaryKey(autoGenerate = true)
+    val id : Int = 0,
+    val name : String = "Training",
+    var isPaid: Boolean,
+    val date : String
+)
