+++ composeApp/src/commonMain/kotlin/database/model/Payment.kt
@@ -1,4 +1,4 @@
-package database
+package database.model
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
